+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 development version
 -------------------
 
+* :issue:`767`: Error messages are always sent to stderr. (They were
+  previously sent to stdout when using `-o`.)
 * Cutadapt can now read single-end data from unaligned BAM files (uBAM).
 * Dropped support for Python 3.7.